+++ src/pages/Home/Home.js
@@ -5,26 +5,9 @@ import { format, addDays, startOfToday } from 'date-fns';
 import { widthPercentageToDP as wp, heightPercentageToDP as hp } from 'react-native-responsive-screen';
 import { useNavigation } from '@react-navigation/native';
 import { useFonts } from "expo-font";
+import Locale from './Local';
+import styles from './HomeStyles';
 
-
-// 캘린더 로케일 설정 분리
-LocaleConfig.locales['kr'] = {
-  monthNames: [
-    '01월', '02월', '03월', '04월', '05월', '06월',
-    '07월', '08월', '09월', '10월', '11월', '12월'
-  ],
-  monthNamesShort: [
-    '01월', '02월', '03월', '04월', '05월', '06월',
-    '07월', '08월', '09월', '10월', '11월', '12월'
-  ],
-  dayNames: [
-    '일요일', '월요일', '화요일', '수요일', '목요일', '금요일', '토요일'
-  ],
-  dayNamesShort: [
-    '일', '월', '화', '수', '목', '금', '토'
-  ],
-};
-LocaleConfig.defaultLocale = 'kr';
 
 const todayString = format(startOfToday(), 'yyyy-MM-dd');
 
@@ -104,11 +87,11 @@ const Home = () => {
     <View style={styles.container}>
       <View style={styles.header}>
         <View style={styles.coin}>
-          <Image source={require('../assets/start.png')} style={styles.star} />
+          <Image source={require('../../assets/start.png')} style={styles.star} />
           <Text style={styles.coinCount}>0</Text>
         </View>
-        <Image source={require('../assets/search.png')} style={styles.search} />
-        <Image source={require('../assets/alarm.png')} style={styles.alarm} />
+        <Image source={require('../../assets/search.png')} style={styles.search} />
+        <Image source={require('../../assets/alarm.png')} style={styles.alarm} />
       </View>
       <View style={styles.body}>
         <Calendar
@@ -154,100 +137,3 @@ const Home = () => {
 
 export default Home;
 
-const styles = StyleSheet.create({
-  container: { flex: 1 },
-  header: {
-    flex: 0.5,
-    flexDirection: 'row',
-    backgroundColor: '#FFFFFF',
-    alignItems: 'flex-start',
-    paddingTop: 50,
-    paddingHorizontal: 20,
-    position: 'relative',
-  },
-  coin: {
-    width: wp('15%'),
-    height: hp('3%'),
-    backgroundColor: '#C8F4ED',
-    borderRadius: 50,
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    flexDirection: 'row',
-    position: 'relative',
-  },
-  star: {
-    width: '55%',
-    height: '90%',
-    resizeMode: 'contain',
-  },
-   coinCount: {
-    position: 'absolute',
-    left: 38,
-    top: 2,
-    fontWeight: '700',
-    fontSize: 20,
-    color: '#454545',
-  },
-  search: {
-    width: wp('5%'),
-    height: hp('3.5%'),
-    resizeMode: 'contain',
-    position: 'absolute',
-    alignItems: 'center',
-    right: 70,
-    top: 50,
-  },
-  alarm: {
-    width: wp('5%'),
-    height: hp('3.5%'),
-    resizeMode: 'contain',
-    position: 'absolute',
-    right: 20,
-    top: 50,
-  },
-  body: {
-    flex: 6.5,
-    backgroundColor: '#FAFAFA',
-    alignItems: 'center',
-  },
-  calendar: {
-    width: wp('90%'),
-    backgroundColor: '#FAFAFA',
-  },
-  journalButton: {
-    padding: 10,
-    backgroundColor: 0,
-    margin: 20,
-    borderRadius: 8,
-    width: wp('80%'),
-    alignSelf: 'center',
-    borderColor: '#ddd',
-    borderTopWidth: 1,
-    borderBottomWidth: 1,
-    borderLeftWidth: 1,
-    borderRightWidth: 1,
-  },
-  journalButtonText: {
-    color: '#000',
-    textAlign: 'center',
-    fontWeight: 'bold',
-  },
-  dayContainer: {
-    backgroundColor: 'white',
-    width: wp('12%'),
-    height: hp('7.5%'),
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  dayInner: {
-    width: 20,
-    height: 14,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  dayText: {
-    fontSize: 10,
-  },
-  enableDate: {},
-  disableDate: { opacity: 0.4 },
-});
