+++ src/pages/Home/Local.js
@@ -1,0 +1,22 @@
+// localeConfig.js
+import { LocaleConfig } from 'react-native-calendars';
+
+LocaleConfig.locales['kr'] = {
+  monthNames: [
+    '01월', '02월', '03월', '04월', '05월', '06월',
+    '07월', '08월', '09월', '10월', '11월', '12월'
+  ],
+  monthNamesShort: [
+    '01월', '02월', '03월', '04월', '05월', '06월',
+    '07월', '08월', '09월', '10월', '11월', '12월'
+  ],
+  dayNames: [
+    '일요일', '월요일', '화요일', '수요일', '목요일', '금요일', '토요일'
+  ],
+  dayNamesShort: [
+    '일', '월', '화', '수', '목', '금', '토'
+  ],
+};
+LocaleConfig.defaultLocale = 'kr';
+
+export default LocaleConfig;
